+++ CHANGELOG.md
@@ -11,11 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Normalize countries to common short names instead of official names, drop unmatched countries, and de-duplicate list
 - Remove "climate change" keyword since it is implied that all journal articles are climate change related and it congests tag clouds and topic lists
+- Normalize publishers
 
 ### Added
 
 - Add column with UN regions (based on countries)
 - Add column continents (based on countries)
+- Retrieve missing publishers from Crossref
 
 ## [1.0.0] - 2024-12-09